requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.103.1
uvicorn>=0.23.2
openai>=1.3.0
gunicorn>=21.2.0
flask>=2.0.0